# docs/swagger/libraryapi_swagger.json
{
    "swagger": "2.0",
    "info": {
...
        "title": "Webtor API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/vault": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Vault Points balance plus every pledge the account holds. 1 VP = 1 GB of torrent size; a `total` of\n`null` means an unlimited plan.\n\nA pledge is frozen for a period after it is created and cannot be removed while it is — `frozen` says\nwhether that applies right now, not whether the content is stored.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vault"
                ],
                "summary": "Vault balance and pledges",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/libapi.VaultResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "402": {
                        "description": "Payment Required",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too many requests with this key — the `Retry-After` header says how long to wait",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Vault is not configured on this deployment",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/vault/pledges": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a pledge, which is what keeps a torrent available long-term. Idempotent per resource: an\naccount holds at most one pledge for a given resource, and pledging again answers `409`.\n\nThe required amount is derived from the torrent size, so a pledge fails with `403` when the account\ndoes not have that many points available.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vault"
                ],
                "summary": "Pledge Vault Points to a torrent",
                "parameters": [
                    {
                        "description": "Resource to pledge to",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/libapi.PledgeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/libapi.Pledge"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "402": {
                        "description": "Payment Required",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Not enough Vault Points",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "A pledge for this resource already exists",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too many requests with this key — the `Retry-After` header says how long to wait",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Vault is not configured on this deployment",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/vault/pledges/{resource_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "One pledge, plus where its transfer stands. Between \"pledged\" and \"vaulted\" this is the only place\nan API client can watch: `status` walks `waiting` → `queued` → `storing` → `vaulted`, with `progress`\n(percent) and `stored_size` / `total_size` (bytes) while a transfer is measurable.\n\n`failed` means the last attempt failed and storage will retry on its own schedule — keep polling,\ndo not re-pledge. `expired` means the resource lost its funding. Poll politely: progress moves on the\norder of tens of seconds, so once per 10–30 seconds is as fresh as it gets.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vault"
                ],
                "summary": "Pledge transfer status",
                "parameters": [
                    {
                        "type": "string",
                        "example": "08ada5a7a6183aae1e09d831df6748d566095a10",
                        "description": "Resource (infohash) the pledge is held against",
                        "name": "resource_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/libapi.PledgeStatusResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "402": {
                        "description": "Payment Required",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "No pledge for this resource",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too many requests with this key — the `Retry-After` header says how long to wait",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "The storage backend did not answer",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Vault is not configured on this deployment",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Removes the pledge for a resource and returns its points to the balance. A pledge that is still\nfrozen cannot be removed and answers `409`.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vault"
                ],
                "summary": "Claim Vault Points back",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Resource (infohash) the pledge is held against",
                        "name": "resource_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Removed"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "402": {
                        "description": "Payment Required",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Pledge is frozen",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too many requests with this key — the `Retry-After` header says how long to wait",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Vault is not configured on this deployment",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "libapi.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "not_found"
                },
                "message": {
                    "type": "string",
                    "example": "the specified path does not exist"
                }
            }
        },
        "libapi.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/libapi.Error"
                }
            }
        },
...
        "libapi.Pledge": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number",
                    "example": 2
                },
                "created_at": {
                    "type": "string",
                    "example": "2026-01-02T15:04:05Z"
                },
                "expired": {
                    "type": "boolean",
                    "example": false
                },
                "frozen": {
                    "description": "Frozen means the points cannot be claimed back yet. It says nothing\nabout whether the content is stored — Vaulted does.",
                    "type": "boolean",
                    "example": true
                },
                "funded": {
                    "type": "boolean",
                    "example": true
                },
                "funded_vp": {
                    "type": "number",
                    "example": 2
                },
                "name": {
                    "type": "string",
                    "example": "Sintel"
                },
                "pledge_id": {
                    "type": "string",
                    "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                },
                "required_vp": {
                    "type": "number",
                    "example": 2
                },
                "resource_id": {
                    "type": "string",
                    "example": "08ada5a7a6183aae1e09d831df6748d566095a10"
                },
                "vaulted": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "libapi.PledgeRequest": {
            "type": "object",
            "required": [
                "resource_id"
            ],
            "properties": {
                "resource_id": {
                    "type": "string",
                    "example": "08ada5a7a6183aae1e09d831df6748d566095a10"
                }
            }
        },
        "libapi.PledgeStatusResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number",
                    "example": 2
                },
                "created_at": {
                    "type": "string",
                    "example": "2026-01-02T15:04:05Z"
                },
                "expired": {
                    "type": "boolean",
                    "example": false
                },
                "frozen": {
                    "description": "Frozen means the points cannot be claimed back yet. It says nothing\nabout whether the content is stored — Vaulted does.",
                    "type": "boolean",
                    "example": true
                },
                "funded": {
                    "type": "boolean",
                    "example": true
                },
                "funded_vp": {
                    "type": "number",
                    "example": 2
                },
                "name": {
                    "type": "string",
                    "example": "Sintel"
                },
                "pledge_id": {
                    "type": "string",
                    "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                },
                "progress": {
                    "type": "number",
                    "x-nullable": true,
                    "example": 42.5
                },
                "required_vp": {
                    "type": "number",
                    "example": 2
                },
                "resource_id": {
                    "type": "string",
                    "example": "08ada5a7a6183aae1e09d831df6748d566095a10"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "waiting",
                        "queued",
                        "storing",
                        "failed",
                        "vaulted",
                        "expired"
                    ],
                    "example": "storing"
                },
                "stored_size": {
                    "type": "integer",
                    "x-nullable": true,
                    "example": 1073741824
                },
                "total_size": {
                    "type": "integer",
                    "x-nullable": true,
                    "example": 2147483648
                },
                "vaulted": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "libapi.VaultContent": {
            "type": "object",
            "properties": {
                "expiring": {
                    "type": "integer",
                    "example": 0
                },
                "loading": {
                    "type": "integer",
                    "example": 1
                },
                "vaulted": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
        "libapi.VaultPoints": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "number",
                    "x-nullable": true,
                    "example": 40
                },
                "claimable": {
                    "type": "number",
                    "example": 40
                },
                "frozen": {
                    "type": "number",
                    "example": 20
                },
                "funded": {
                    "type": "number",
                    "example": 60
                },
                "total": {
                    "type": "number",
                    "x-nullable": true,
                    "example": 100
                }
            }
        },
        "libapi.VaultResponse": {
            "type": "object",
            "properties": {
                "content": {
                    "$ref": "#/definitions/libapi.VaultContent"
                },
                "pledges": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/libapi.Pledge"
                    }
                },
                "points": {
                    "$ref": "#/definitions/libapi.VaultPoints"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "API key from your profile page, as `Bearer \u003ckey\u003e`.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}